% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{aa_las_chablais3}
\alias{aa_las_chablais3}
\title{Load las_chablais3}
\usage{
aa_las_chablais3()
}
\value{
An object of class \code{\link[lidR]{LAS}}.
}
\description{
Loads the external data: Airborne laser scanning data over the
 Chablais 3 plot, acquired in 2009 by Sintegra
}
\examples{
las_chablais3 <- aa_las_chablais3()
}
\seealso{
\code{\link{las_chablais3}}
}
\keyword{internal}
