% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba_inference}
\alias{aba_inference}
\title{computes inference from area-based model and predicted values}
\usage{
aba_inference(
  aba_model,
  r_predictions,
  type = c("SRS", "ED", "D", "STR", "SYNT"),
  r_mask = NULL
)
}
\arguments{
\item{aba_model}{a model returned by \code{\link{aba_build_model}} or
\code{\link{aba_combine_strata}}}

\item{r_predictions}{raster of predicted values}

\item{type}{string vector specifying which estimators should be computed
(one or several in "SRS", "ED", "D", "STR", "SYNT")}

\item{r_mask}{raster to mask region of interest (NA values), may contain
post-stratification categories (should be integer, positive values)}
}
\value{
a data frame with estimation of parameter value and standard deviation
 of estimation for all required estimators.
}
\description{
computes inference from area-based model and predicted values
}
