% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{aba_metrics}
\alias{aba_metrics}
\alias{.aba_metrics}
\title{Function for area-based metrics computation}
\format{
An object of class \code{formula} of length 2.
}
\usage{
aba_metrics(z, i, rn, c, hmin = 2, breaksH = NULL)

.aba_metrics
}
\arguments{
\item{z, i, rn, c}{Height, Intensity, ReturnNumber and Classification}

\item{hmin}{numeric. height threshold for low points removal before metrics
computation}

\item{breaksH}{vector. breaks for height histogram proportion computation}
}
\description{
Predefined function usable in \code{\link[lidR]{cloud_metrics}} or
\code{\link{clouds_metrics}}. Applies a minimum height threshold to the point
cloud and computes the following metrics:
\enumerate{
\item for all points: total number \code{ntot}, percentage of points above
minimum height \code{p_hmin}, percentage of points in height bins
\code{H.propZ1_Z2},
\item for first return points: percentage above minimum height
\code{p_1st_hmin},
\item for all points above minimum height: height metrics returned by
\code{\link[lidR]{stdmetrics_z}} and intensity metrics returned by
\code{\link[lidR]{stdmetrics_i}}
\item for first returns above minimum height: \code{mCH} and \code{sdCH} as
proposed by Bouvier et al.
}
}
\examples{
# load LAS file
LASfile <- system.file("extdata", "las_chablais3.laz", package="lidaRtRee")
las_chablais3 <- lidR::readLAS(LASfile)

# set number of threads
lidR::set_lidr_threads(2)
# extract two point clouds from LAS object
llas <- lidR::clip_circle(las_chablais3,
                          c(974350, 974390),
                          c(6581680, 6581680), 10)
# normalize point clouds
llas <- lapply(llas, function(x) {
  lidR::normalize_height(x, lidR::tin())
})
# computes metrics 
m <- clouds_metrics(llas, ~ aba_metrics(
 Z, Intensity, ReturnNumber, Classification, 2
))
head(m[,1:5])
}
\references{
Bouvier et al. 2015. Generalizing predictive models of forest
inventory attributes using an area-based approach with airborne LiDAR data.
Remote Sensing of Environment 156, pp. 322-334. \doi{10.1016/j.rse.2014.10.004}
}
\seealso{
\code{\link[lidR]{cloud_metrics}}, \code{\link[lidR]{stdmetrics}},
\code{\link{clouds_metrics}}
}
\keyword{datasets}
