% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{ABAinference}
\alias{ABAinference}
\title{computes inference from area-based model and predicted values}
\usage{
ABAinference(
  modell,
  r.predictions,
  type = c("SRS", "ED", "D", "STR", "SYNT"),
  r.mask = NULL
)
}
\arguments{
\item{modell}{a model returned by by \code{\link{ABAmodel}} or \code{\link{ABAmodelCombineStrata}}}

\item{r.predictions}{raster of predicted values}

\item{type}{string vector specifying which estimators should be computed (one or several in "SRS", "ED", "D", "STR", "SYNT")}

\item{r.mask}{raster to mask region of interest (NA values), may contain post-stratification categories (should be integer, positive values)}
}
\value{
a dataframe with estimation of parameter value and standard deviation of estimation for all required estimators.
}
\description{
computes inference from area-based model and predicted values
}
