% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chm_chablais3-data.R
\docType{data}
\name{chm_chablais3}
\alias{chm_chablais3}
\title{Canopy height model (Chablais 3 plot)}
\format{
A PackedSpatRaster object
}
\usage{
data(chm_chablais3)
}
\description{
Canopy height model computed from airborne laser scanning data acquired in July 2010.
}
\examples{
data(chm_chablais3)
chm_chablais3 <- terra::rast(chm_chablais3)
terra::plot(chm_chablais3)
}
\references{
Monnet, J.-M. 2011. Using airborne laser scanning for mountain forests mapping: Support vector regression for stand parameters estimation and unsupervised training for treetop detection. Ph.D. thesis. University of Grenoble, France. pp. 21-22 & 34 \url{https://theses.hal.science/tel-00652698/document}
}
\keyword{datasets}
