% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chmchablais3-data.R
\docType{data}
\name{chmchablais3}
\alias{chmchablais3}
\title{Canopy height model (Chablais 3 plot)}
\format{
A raster object
}
\usage{
data(chmchablais3)
}
\description{
Canopy height model computed from airborne laser scanning data acquired in July 2010.
}
\examples{
data(chmchablais3)
chmchablais3
raster::plot(chmchablais3)
}
\references{
Monnet, J.-M. 2011. Using airborne laser scanning for mountain forests mapping: Support vector regression for stand parameters estimation and unsupervised training for treetop detection. Ph.D. thesis. University of Grenoble, France. pp. 21-22 & 34 \url{https://tel.archives-ouvertes.fr/tel-00652698/document}
}
\keyword{datasets}
