% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.detection.R
\name{createDisk}
\alias{createDisk}
\title{Disk-shaped matrix mask}
\usage{
createDisk(width = 5)
}
\arguments{
\item{width}{numeric. disk width in pixels, should be an uneven number}
}
\value{
A matrix with 1 for pixels inside the disk, 0 outside
}
\description{
Creates a matrix with TRUE values shaping a centered disk
}
\examples{
createDisk(7)
}
