% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_match.R
\name{hist_stack}
\alias{hist_stack}
\title{Stacked histogram}
\usage{
hist_stack(x, breaks, col = NULL, breaksFun = paste, ...)
}
\arguments{
\item{x}{list of vectors. values for each category}

\item{breaks}{vector. breaks for histogram bins}

\item{col}{vector. colors for each category}

\item{breaksFun}{function for breaks display}

\item{...}{arguments to be passed to methods, as in \code{\link[graphics]{plot}}}
}
\value{
no return
}
\description{
Stacked histogram
}
