% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laschablais3-data.R
\docType{data}
\name{laschablais3}
\alias{laschablais3}
\title{las data in France (Chablais 3 plot)}
\format{
An object of class \code{\link[lidR]{LAS}}
}
\source{
Monnet J.-M. INRAE
}
\usage{
data(laschablais3)
}
\description{
Airborne laser scanning data over the Chablais 3 plot, acquired in 2009 by Sintegra, copyright INRAE
}
\details{
Additional information about the data
\itemize{
\item{Sensor: RIEGL LMS-Q560)}
\item{EPSG code of coordinates system: 2154}
}
}
\examples{
data(laschablais3)
laschablais3
# display point cloud
lidR::plot(laschablais3)
}
\references{
Monnet, J.-M. 2011. Using airborne laser scanning for mountain forests mapping: Support vector regression for stand parameters estimation and unsupervised training for treetop detection. Ph.D. thesis. University of Grenoble, France. pp. 21-22. \url{https://tel.archives-ouvertes.fr/tel-00652698/document}
}
\keyword{datasets}
