% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{rasterChullMask}
\alias{rasterChullMask}
\title{Raster mask of convex hull}
\usage{
rasterChullMask(xy, r)
}
\arguments{
\item{xy}{2 columns matrix or data.frame. xy positions}

\item{r}{raster object. target raster}
}
\value{
a raster with 0 or 1
}
\description{
creates raster mask corresponding to the convex hull of xy positions
}
\examples{
# create raster
r <- raster::raster()
raster::extent(r) <- c(0,40,0,40)
raster::res(r) <- 1

# xy positions
xy <- data.frame(c(10,20,31.25,15),
  c(10,20,31.25,25))
# compute mask
mask1 <- rasterChullMask(xy, r)

# display binary raster
raster::plot(mask1)
graphics::points(xy)
}
\seealso{
\code{\link{rasterXYMask}}
}
