% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{species_color}
\alias{species_color}
\title{Table of species names, abreviations and display colors}
\usage{
species_color()
}
\value{
A data frame with species name, color, coniferous (C) / broadleaf (B) 
type, and name abreviation GESP of GEnus and SPecies
}
\description{
table for species names, abreviations and type (coniferous/broadleaf), and 
display color
}
\examples{
# load table
tab.species <- species_color()
head(tab.species)
summary(tab.species)
}
\seealso{
\code{\link{plot_tree_inventory}} for tree inventory display
}
