% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{speciesColor}
\alias{speciesColor}
\title{Table of species names, abreviations and display colors}
\usage{
speciesColor()
}
\value{
A data frame with species name, color, coniferous (C) / broadleaf (B) type, and name abreviation GESP of GEnus and SPecies
}
\description{
table for species names, abreviations and type (coniferous/broadleaf), and display color
}
\examples{
# load table
tab.species <- speciesColor()
head(tab.species)
summary(tab.species)
}
\seealso{
\code{\link{plotTreeInventory}} for tree inventory display
}
