% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_detection.R
\name{maxima_selection}
\alias{maxima_selection}
\title{Image maxima selection based on values and neighborhood of local maxima}
\usage{
maxima_selection(maxi, dem_nl, hmin = 5, dmin = 0, dprop = 0.05)
}
\arguments{
\item{maxi}{cimg object or SpatRaster object. image with local maxima 
(typically output from \code{\link{maxima_detection}}, image values correspond 
to neighborhood radius on which pixels are global maxima in the initial image)}

\item{dem_nl}{cimg object or SpatRaster object. initial image from which maxima were detected}

\item{hmin}{numeric. minimum value in initial image for a maximum to be selected}

\item{dmin}{numeric. intercept term for selection of maxima depending on 
neighborhood radius: \code{maxi >= dmin + dem_nl * dprop}}

\item{dprop}{numeric. proportional term for selection of maxima depending on 
neighborhood radius: \code{maxi >= dmin + dem_nl * dprop}}
}
\value{
A cimg object or SpatRaster object which values are the radius (n) 
in meter of the square window (width 2n+1) where the center pixel is global 
maximum and which fulfill the selection criteria
}
\description{
In a maxima image (output of \code{\link{maxima_detection}}), sets values to 
zero for pixels which:
\enumerate{
\item values in the initial image (from which maxima were detected) are below 
a threshold
\item values in the maxima image (corresponding to the radius of the 
neighborhood where they are global maxima) are below a threshold depending on 
the initial image value.
}
Make sure that the \code{max.width} parameter in \code{\link{maxima_detection}}
is consistent with the selection parameters (e.g. do not select with 
\code{dmin = 7} if values were only tested up to \code{max.width} the default 
value which is approx. 5.5 m).
}
\examples{
data(chm_chablais3)
chm_chablais3 <- terra::rast(chm_chablais3)

# maxima detection
maxi <- maxima_detection(chm_chablais3)

# several maxima selection settings
selected_maxi_hmin <- maxima_selection(maxi, chm_chablais3, hmin = 15)
selected_maxi_dm <- maxima_selection(maxi, chm_chablais3, dmin = 2.5)
selected_maxi <- maxima_selection(maxi, chm_chablais3, dmin = 1, dprop = 0.1)

# corresponding count number of remaining maxima
table(terra::values(maxi))
table(terra::values(selected_maxi_hmin))
table(terra::values(selected_maxi_dm))
table(terra::values(selected_maxi))

# plot original image
terra::plot(chm_chablais3, main = "Initial image")

# plot maxima images, original and first case
terra::plot(maxi, main = "Local maxima")
terra::plot(selected_maxi, main = "Selected maxima")
}
\seealso{
\code{\link{maxima_detection}}, \code{\link{tree_segmentation}}
}
